몹 바닥으로 떨어지는거 방지

+++ client/UnityClient/Assets/Scripts/Tile/MapData.cs
@@ -18,7 +18,7 @@ namespace TileData
 
     public class Layer
     {
-        public int[] Data { get; set; }
+        public short[] Data { get; set; }
 
         public string Name { get; set; }
